requirements.txt
discord.py
wynncraft